Add option to frig roundabouts.
Add option to "frig" roundabouts by adding an extra point between each
pair of adjacent nodes. The extra points are positioned "outside" of
the line joining the two nodes. The effect is to make roundabouts
"rounder". Unfortunately, it also makes them ugly as the frigging is
being done by a dumb computer with little aesthetic capability.

It is not enabled by default, you need to specify the --frig-roundabouts
option. By default the "frig factor" (influences how far the new points are
away from the centre of the roundabout) is set to 0.25. You can specify a
different value as the option value: --frig-roundabouts=0.1

Large values tend to make the roundabout uglier so it would be worth
experimenting to see what the minimum frig factor is that will stop the
gps from reporting the wrong exit.

You can also enable frigging for an individual roundabout by tagging the OSM
XML with the "mkgmap:frig_roundabout" tag which can either be set to a valid
(floating point) number to specify the frig factor or just some other string
that will make it use the default frig factor.

This patch also avoids putting in new points if the new point would be
less than 5 or more than 100 metres from the adjacent points.

- Mark Burton



git-svn-id: http://svn.parabola.me.uk/mkgmap/branches/nod@881 25d90789-57f7-4ee0-8453-03a3dfeeeb22