Fix rounding in two places.
Before the patch clipping was a little inaccurate. Some points 2m outside
the bounding box gets included in the img. With this patch it works
correctly now.

I've tested this since some days and up to now found no new failures.
Whatever the reason for this small delta was, there should be no bad
effects, if one adds a somewhat bigger delta.

- Johann Gail

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@996 25d90789-57f7-4ee0-8453-03a3dfeeeb22