Avoid making unwanted nodes when --link-pois-to-ways is active.
When linking pois to ways it makes sense to consider the poi to be a
node for the purposes of short arc removal. Unfortunately, this was done
in a buggy fashion that would cause nodes to be created when they should
not have been. Now, it just notices that a point could possibly become a
node but it no longer increments the point's highway count so it won't
become a node unless it really is shared by more than one way.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1221 25d90789-57f7-4ee0-8453-03a3dfeeeb22