Correct the map set block.
There is one value in this block and this item is the only thing that does not come from
the tdb or overview map.  So I have added a new option --mapset-name for this.  In mapsource
there is a field you can fill in before sending to the device labeled Map Set Name.  This is the same
thing.

git-svn-id: http://svn.parabola.me.uk/mkgmap/branches/mapset@1325 25d90789-57f7-4ee0-8453-03a3dfeeeb22