Create branch for optimising the new style engine.
The performance is fine for the default style (no different to before)
but somewhat worse for the openmtbmap style.  Everyone else is
somewhere in between.

Previously there was an index so that you only run rules that might
match.  Now we run all rules until we find a match.  This branch
will explore ways of resticting the number of attempted matches.


git-svn-id: http://svn.parabola.me.uk/mkgmap/branches/style-speed@1550 25d90789-57f7-4ee0-8453-03a3dfeeeb22