This patch enables mkgmap to make use of real multipolygons and of multipolygons where only the members should be tagged (e.g. country boundaries) without any change on the style rules.
It adds a copy of each original line of a multipolygon with an outer role, tags them with the mp tags and an additional mkgmap:stylefilter=polyline. The generated polygon elements are tagged with mkgmap:stylefilter=polygon. The style system evaluates the mkgmap:stylefilter tag and processes such tagged elements only with the appropriate style rules.

The --process-boundary-relations option is no longer neccessary and has been removed.

- WanMil

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1694 25d90789-57f7-4ee0-8453-03a3dfeeeb22