Only apply merge lines at lower resolutions.
merge-lines is as discussed quite problematic, because it happens after the routing notes are placed, and therefore causes some problems. 
The closer you zoom in, the bigger these problems, while the further you zoom out, the bigger the speed improvement of using this
option (at resolution 24 and 23 it makes basically no difference) and the smaller the problems (22 vs 24 means 1/4 of the wrong distance
for info popup or jumping map when panning). 

This only has any affect if you are using the --merge-lines option.

- Felix Hartmann

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@1876 25d90789-57f7-4ee0-8453-03a3dfeeeb22