Cities with the same details but different map tiles are both recorded in the index.
This is because when you do not select the tile that the city name happens to be
in then it cannot be found and gives an error on transfer to device. I assume so
anyway, my own mdrcheck tools gives an error for that reason.

Unfortunately this brings back the previous mdr_trim_addr error.

git-svn-id: http://svn.parabola.me.uk/mkgmap/branches/city-region-index@1991 25d90789-57f7-4ee0-8453-03a3dfeeeb22