Boundary precompiler uses a finer detection if a boundary is usable
The detection if a boundary is usable now checks also if a name tag (a tagname containing the string 'name') is set which is mandatory to be used by the LocationHook. Otherwise the boundaries blow up the precompiled bounds but must be sorted out by the LocationHook. When doing this in the precompilation step it saves time while generating the real map.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@2051 25d90789-57f7-4ee0-8453-03a3dfeeeb22