Fix rare NullPointerException when loading preprocessed boundary
The NullPointerException happened when only a very small area of the boundary intersects the bounds bounding box. In rare cases cutting the area to the bbox leaves one small area only. Fitting this area to the integer coord space changes the direction from clockwise to counter clockwise which mangles the detection if an area is outer or inner. This misdetection causes the NullPointerException.

git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@2202 25d90789-57f7-4ee0-8453-03a3dfeeeb22