Merge from short-lived 2gb branch.
Deals with files greater than 2GB in size.  Also correct the partition
sizes.  First pick a "disk" size from a fixed list of possibilities;
the first one that is larger than the file.  Then set the partition
size to be just big enough to hold the file.

It seems that new devices (dakota, montana etc) care about this
stuff, older devices do not. Mapsource/Basecamp has always been
sensitive to this being right.

I'm not entirely sure of the exact rules. In general the partition size
just needs to be larger than the file, but there were problems in the
past when it was very much bigger. On the other hand those problems
occurred when there were other faults, so it could be they were the
real problem.


git-svn-id: http://svn.parabola.me.uk/mkgmap/trunk@2261 25d90789-57f7-4ee0-8453-03a3dfeeeb22